+++ packages/fiori-app-sub-generator/test/int/fiori-elements/expected-output/alp_v4/webapp/test/integration/FirstJourney.js
@@ -8,6 +8,8 @@ sap.ui.define([
         QUnit.module("First journey");
 
         opaTest("Start application", function (Given, When, Then) {
+            Given.iResetMockData({ ServiceUri: "/sap/opu/odata4/sap/c_salesordermanage_srv/srvd/sap/c_salesordermanage_sd_aggregate/0001/" });
+            Given.iResetTestData();
             Given.iStartMyApp();
             
         });