+++ packages/deploy-config-generator-shared/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sap-ux/deploy-config-generator-shared
 
+## 1.0.40
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/fiori-generator-shared 1.2.17 → 1.2.18
+
+## 1.0.39
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/fiori-generator-shared 1.2.16 → 1.2.17
+
 ## 1.0.38
 
 ### Patch Changes