+++ packages/preview-middleware/CHANGELOG.md
@@ -1,5 +1,56 @@
 # @sap-ux/preview-middleware
 
+## 1.2.19
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/adp-tooling 1.0.58 → 1.0.59
+
+## 1.2.18
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/i18n 1.0.3 → 1.0.4
+- @sap-ux/adp-tooling 1.0.57 → 1.0.58
+- @sap-ux/project-access 2.2.0 → 2.2.1
+- @sap-ux/system-access 1.0.14 → 1.0.14
+
+## 1.2.17
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/adp-tooling 1.0.56 → 1.0.57
+
+## 1.2.16
+
+### Patch Changes
+
+#### Release Date
+
+2026-09-21
+
+#### Features
+
+- Serve merged manifest.json via proxy in CF ADP preview mode [[3f822e1](https://github.com/SAP/open-ux-tools/commit/3f822e140aad6deb2df243295a25a917b7044929)]
+
+#### Workspace Updates
+
+- @sap-ux/adp-tooling 1.0.55 → 1.0.56
+
+## 1.2.15
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/adp-tooling 1.0.54 → 1.0.55
+
 ## 1.2.14
 
 ### Patch Changes