+++ packages/ui5-library-reference-sub-generator/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @sap-ux/ui5-library-reference-sub-generator
 
+## 1.0.45
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/project-access 2.2.0 → 2.2.1
+- @sap-ux/fiori-generator-shared 1.2.17 → 1.2.18
+- @sap-ux/telemetry 2.0.7 → 2.0.8
+- @sap-ux/ui5-library-reference-inquirer 1.0.44 → 1.0.45
+- @sap-ux/ui5-library-reference-writer 1.0.19 → 1.0.20
+
+## 1.0.44
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/telemetry 2.0.6 → 2.0.7
+- @sap-ux/fiori-generator-shared 1.2.16 → 1.2.17
+- @sap-ux/ui5-library-reference-inquirer 1.0.43 → 1.0.44
+
+## 1.0.43
+
+### Patch Changes
+
+#### Workspace Updates
+
+- @sap-ux/ui5-library-reference-inquirer 1.0.42 → 1.0.43
+
 ## 1.0.42
 
 ### Patch Changes